Split simpleNbox.cpp into two files
It was 1300+ lines long!
